requirements.txt
absl-py==0.8.1
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
bleach==1.5.0
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
fancycompleter==0.8
future==0.18.2
gast==0.3.2
grpcio==1.25.0
h5py==2.10.0
html5lib==0.9999999
idna==2.8
importlib-metadata==0.23
imutils==0.5.3
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.17.4
opencv-contrib-python==4.1.2.30
pandocfilters==1.4.2
parso==0.5.1
pdbpp==0.10.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.0
ptyprocess==0.6.0
Pygments==2.5.1
pyparsing==2.4.5
pyrsistent==0.15.6
python-dateutil==2.8.1
PyYAML==5.1.2
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
scipy==1.3.3
Send2Trash==1.5.0
six==1.13.0
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
tqdm==4.40.0
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wmctrl==0.3
zipp==0.6.0
